# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Recipes and Health Tips API",
    "description": "API for managing health tips and recipes with CRUD operations",
    "version": "1.0.0"
  },
  "host": "cse341-project2-z10v.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["Hello World"],
        "description": "Welcome endpoint",
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/healthtips/": {
      "get": {
        "tags": ["healthtips"],
        "description": "Get all health tips",
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "Successfully retrieved all health tips" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["healthtips"],
        "description": "Create a new health tip",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Health tip to be created",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["title", "description", "category"],
              "properties": {
                "title": { "type": "string", "example": "Drink Water" },
                "description": { "type": "string", "example": "Stay hydrated by drinking 8 glasses of water daily." },
                "category": { "type": "string", "example": "Nutrition" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Health tip created successfully" },
          "400": { "description": "Validation failed" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/healthtips/{id}": {
      "get": {
        "tags": ["healthtips"],
        "description": "Get a single health tip by ID",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Health tip ID"
          }
        ],
        "responses": {
          "200": { "description": "Successfully retrieved health tip" },
          "400": { "description": "Invalid ID format" },
          "404": { "description": "Health tip not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["healthtips"],
        "description": "Update a health tip by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Health tip ID"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated health tip object",
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string", "example": "Drink More Water" },
                "description": { "type": "string", "example": "Aim for 10 glasses of water daily." },
                "category": { "type": "string", "example": "Hydration" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Health tip updated successfully" },
          "400": { "description": "Validation failed or invalid ID" },
          "404": { "description": "Health tip not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["healthtips"],
        "description": "Delete a health tip by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Health tip ID"
          }
        ],
        "responses": {
          "200": { "description": "Health tip deleted successfully" },
          "400": { "description": "Invalid ID format" },
          "404": { "description": "Health tip not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/recipes/": {
      "get": {
        "tags": ["recipes"],
        "description": "Get all recipes",
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "Successfully retrieved all recipes" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["recipes"],
        "description": "Create a new recipe",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Recipe object to be created",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "ingredients", "instructions"],
              "properties": {
                "name": { "type": "string", "example": "Grilled Chicken Salad" },
                "ingredients": { "type": "array", "items": { "type": "string" }, "example": ["Chicken", "Lettuce", "Tomatoes", "Dressing"] },
                "instructions": { "type": "string", "example": "Grill the chicken, chop vegetables, and mix all together with dressing." }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Recipe created successfully" },
          "400": { "description": "Validation failed" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  ,
    "/recipes/{id}": {
      "get": {
        "tags": ["recipes"],
        "description": "Get a single recipe by ID",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Recipe ID"
          }
        ],
        "responses": {
          "200": { "description": "Successfully retrieved recipe" },
          "400": { "description": "Invalid ID format" },
          "404": { "description": "Recipe not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["recipes"],
        "description": "Update a recipe by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Recipe ID"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated recipe object",
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "example": "Updated Grilled Chicken Salad" },
                "ingredients": { 
                  "type": "array", 
                  "items": { 
                    "$ref": "#/definitions/Ingredient" 
                  }, 
                  "example": ["Chicken", 
                            {"name":"Lettuce", 
                            "_id":"60c72b2f9b1e8b001c8e4d1a"}] 
                },
                "instructions": { "type": "string", "example": "Grill the chicken, chop vegetables, and mix all together with dressing." }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Recipe updated successfully" },
          "400": { "description": "Validation failed or invalid ID" },
          "404": { "description": "Recipe not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["recipes"],
        "description": "Delete a recipe by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Recipe ID"
          }
        ],
        "responses": {
          "200": { "description": "Recipe deleted successfully" },
          "400": { "description": "Invalid ID format" },
          "404": { "description": "Recipe not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  }
}